requirements.txt
pandas
requests
gspread
oauth2client
python-dotenv
pyyaml
